+++ src/main/java/rbasamoyai/createbigcannons/cannons/autocannon/AutocannonBlock.java
@@ -57,7 +57,7 @@ public interface AutocannonBlock extends WeldableBlock, CannonContraptionProvide
                     && be1 instanceof IAutocannonBlockEntity cbe1) {
                 cbe1.cannonBehavior().setConnectedFace(opposite, false);
                 cbe1.cannonBehavior().setWelded(opposite, false);
-                be1.setChanged();
+				cbe1.cannonBehavior().blockEntity.setChanged();
             }
         }
 
@@ -72,7 +72,7 @@ public interface AutocannonBlock extends WeldableBlock, CannonContraptionProvide
                     && be2 instanceof IAutocannonBlockEntity cbe2) {
                 cbe2.cannonBehavior().setConnectedFace(facing, false);
                 cbe2.cannonBehavior().setWelded(facing, false);
-                be2.setChanged();
+                cbe2.cannonBehavior().blockEntity.setChanged();
             }
         }
     }
@@ -106,7 +106,7 @@ public interface AutocannonBlock extends WeldableBlock, CannonContraptionProvide
             if (level.getBlockEntity(pos1) instanceof IAutocannonBlockEntity cbe1) {
                 cbe.cannonBehavior().setConnectedFace(facing, true);
                 cbe1.cannonBehavior().setConnectedFace(facing.getOpposite(), true);
-                be1.setChanged();
+				cbe1.cannonBehavior().blockEntity.setChanged();
                 Vec3 particlePos = center.add(offset);
                 slevel.sendParticles(ParticleTypes.CRIT, particlePos.x, particlePos.y, particlePos.z, 10, 0.5d, 0.5d, 0.5d, 0.1d);
             }
@@ -126,7 +126,7 @@ public interface AutocannonBlock extends WeldableBlock, CannonContraptionProvide
             if (level.getBlockEntity(pos2) instanceof IAutocannonBlockEntity cbe2) {
                 cbe.cannonBehavior().setConnectedFace(facing.getOpposite(), true);
                 cbe2.cannonBehavior().setConnectedFace(facing, true);
-                be2.setChanged();
+				cbe2.cannonBehavior().blockEntity.setChanged();
                 Vec3 particlePos = center.add(offset.reverse());
                 slevel.sendParticles(ParticleTypes.CRIT, particlePos.x, particlePos.y, particlePos.z, 10, 0.5d, 0.5d, 0.5d, 0.1d);
             }