+++ src/main/java/rbasamoyai/createbigcannons/cannons/big_cannons/BuiltUpCannonCTBehavior.java
@@ -73,4 +73,5 @@ public class BuiltUpCannonCTBehavior extends ConnectedTextureBehaviour.Base {
 		}
 		return ctx;
 	}
+
 }